requirements.txt
requests==2.28.1
PyTelegramBotAPI==4.8.0
telebot~=0.0.4
timezonefinder~=6.2.0

# Core dependencies
python-telegram-bot>=20.8
requests>=2.31.0
urllib3>=2.0.0
timezonefinder>=6.2.0
pytz>=2023.3
charset-normalizer>=3.0.0

# Optional but recommended
types-requests>=2.31.0    # Type hints for requests library
mypy>=1.8.0              # Static type checker
black>=24.3.0           # Code formatter
pylint>=3.0.3            # Code linter

# Development tools (optional)
pytest>=7.4.4            # For testing
python-dotenv>=1.0.0     # For environment variable management

